requirements.txt
ipaddress==1.0.23
mysql-connector-python==9.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
pyodbc==5.2.0